refactor: Create methods/ in annotations
