+++ src/winter/annotations/methods/RequestParam.java
@@ -1,4 +1,4 @@
-package winter.annotations;
+package winter.annotations.methods;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -6,7 +6,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.ElementType;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface UrlMapping {
-    String value() default "";
+@Target(ElementType.PARAMETER)
+public @interface RequestParam {
+    String name() default "";
 }
