+++ src/winter/annotations/methods/Rest.java
@@ -1,4 +1,4 @@
-package winter.annotations;
+package winter.annotations.methods;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -6,7 +6,6 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.ElementType;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.PARAMETER)
-public @interface RequestParam {
-    String name() default "";
+@Target(ElementType.METHOD)
+public @interface Rest {
 }
