+++ src/winter/annotations/methods/UrlMapping.java
@@ -1,4 +1,4 @@
-package winter.annotations;
+package winter.annotations.methods;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -7,5 +7,6 @@ import java.lang.annotation.ElementType;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
-public @interface Rest {
+public @interface UrlMapping {
+    String value() default "";
 }
