+++ src/winter/utils/AnnotationScanner.java
@@ -16,6 +16,7 @@ import winter.exceptions.InvalidPackageNameException;
 import winter.exceptions.PackageProviderNotFoundException;
 import winter.annotations.*;
 import winter.annotations.classes.Controller;
+import winter.annotations.methods.UrlMapping;
 
 public class AnnotationScanner extends Utility {
     public static Map<String, Mapping> scanControllers(ServletContext servletContext, Map<String, Mapping> urlMappings)
